Improving selection of tools
- Now tool can be activated/deactivated with animations. Tools that
aren't active are hidden after the animation.
- In order to create the moving transition I changed the ToolIcon class
so that it has 2 rectangles, one that contains the icon itself that is
drawn and that is put as child of the panel, and another one that is
used for raytracing and isn't child of the panel but its position is
computed at each frame, to avoid the problem that IsColliding() doesn't
work with objects that have a parent.
- Now only one panel at a time can be activated (i.e. if I have a tool
selected in one panel and then select a tool in another one, the
previous one is disabled)