Code cleaning complete
- Finished restructuring some of the code, cleanining it and commenting
it.

The structure now it's like this:

Scene (spawns both controllers) ---> InteractionManager (uses both
controllers) ---> ToolsMenu ---> Panels ---> Tools

- I found out a design problem: with the code like this it's very hard
to modifiy external stuff, like enabling and disabiling the Skybox. It's
not very clean to just pass it to the interaction manager, and then to
the tools menu, and then to each panel just to pass it to the tool that
actually needs it.

So, for now at least, I made the Skybox a global variable so that the
tool that uses it can directly access it. Not the best thing.