Testing with the "clipping tool" function (it's not in the code anymore, but there is the idea)
I tried using a stencil buffer but it doesn't work as I thought. The
only option left (that I know of) was to use a shader and discard the
fragments above the given plane, and this seems to work.

In the code there is a test shader that discards fragments with y above
the given value; I tried with a VRP_TUBE object and it worked. The
problem is that for a given mesh I need a texture or something so that I
can draw their actual color from the fragment shader (in the test I
simply draw red the good pixels and discarded the others).

Also there is to consider the fact that for this to work for any mesh
that might use any shader, the code should be in any shader.

To make it work with the shader we have to draw back and front faces,
like this:

testShader.Start();
testShader.SetParameter("y", mActionController.GetPosition().y);

glCullFace(GL_FRONT);
mObject.Draw();
glCullFace(GL_BaCK);
mObject.Draw();
testShader.Stop();